Add support to build the upstream llvm master trunk.  Set AOMP_BUILD_TRUNK=1 to turn this on. The trunk will use a different set of patches, hopefully smaller set

+++ bin/build_openmp.sh
@@ -129,6 +129,9 @@ fi
 
 # This is how we tell the hsa plugin where to find hsa
 export HSA_RUNTIME_PATH=$ROCM_DIR/hsa
+if [ "$AOMP_BUILD_TRUNK" != 0 ] ; then
+   export HIP_DEVICE_LIB_PATH=$ROCM_DIR/lib
+fi
 
 if [ "$1" != "nocmake" ] && [ "$1" != "install" ] ; then 
 
@@ -160,6 +163,8 @@ if [ "$1" != "nocmake" ] && [ "$1" != "install" ] ; then
          exit 1
       fi
 
+   # We cannot build debug libs with hip till hip has support for robust posix printf
+   if [ "$AOMP_BUILD_TRUNK" != 0 ] ; then
       echo rm -rf $BUILD_DIR/build/openmp_debug
       rm -rf $BUILD_DIR/build/openmp_debug
       MYCMAKEOPTS="$COMMON_CMAKE_OPTS -DLIBOMPTARGET_NVPTX_DEBUG=ON -DCMAKE_BUILD_TYPE=Debug $AOMP_ORIGIN_RPATH -DROCM_DIR=$ROCM_DIR"
@@ -179,6 +184,7 @@ if [ "$1" != "nocmake" ] && [ "$1" != "install" ] ; then
          echo "      $MYCMAKEOPTS"
          exit 1
       fi
+   fi
 fi
 
 cd $BUILD_DIR/build/openmp
@@ -194,6 +200,7 @@ if [ $? != 0 ] ; then
       exit 1
 fi
 
+if [ "$AOMP_BUILD_TRUNK" != 0 ] ; then
 cd $BUILD_DIR/build/openmp_debug
 echo
 echo
@@ -211,6 +218,7 @@ else
       echo
   fi
 fi
+fi
 
 #  ----------- Install only if asked  ----------------------------
 if [ "$1" == "install" ] ; then 
@@ -222,6 +230,8 @@ if [ "$1" == "install" ] ; then
          echo "ERROR make install failed "
          exit 1
       fi
+
+      if [ "$AOMP_BUILD_TRUNK" != 0 ] ; then
       cd $BUILD_DIR/build/openmp_debug
       echo
       echo " -----Installing to $INSTALL_OPENMP/lib-debug ---- " 
@@ -229,5 +239,6 @@ if [ "$1" == "install" ] ; then
       if [ $? != 0 ] ; then 
          echo "ERROR make install failed "
          exit 1
+      fi
       fi
 fi
